#8 | #9 | Métodos de criação e busca para o usuário, mapeamento dos enums

+++ WebLuto/Models/User.cs
@@ -8,7 +8,7 @@ namespace WebLuto.Models
 
         public string Password { get; set; }
 
-        public int Salt { get; set; }
+        public long Salt { get; set; }
 
         public UserType Type { get; set; }
     }
