+++ WebLuto/Security/Sha512Cryptographer.cs
@@ -5,7 +5,7 @@ namespace WebLuto.Security
 {
     public static class Sha512Cryptographer
     {
-        public static string Encrypt(string text, int salt)
+        public static string Encrypt(string text, long salt)
         {
             var password = Encoding.UTF8.GetBytes(text);
             byte[] saltBytes = Encoding.ASCII.GetBytes(salt.ToString());
@@ -15,7 +15,7 @@ namespace WebLuto.Security
             return Convert.ToBase64String(saltedHash);
         }
 
-        public static bool Compare(string plainValue, int salt, string hashedValue)
+        public static bool Compare(string plainValue, long salt, string hashedValue)
         {
             return Encrypt(plainValue, salt).Equals(hashedValue);
         }
